% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RptCfg}
\alias{RptCfg}
\title{NCA Report Configuation Table}
\format{A data frame with 48 observations on the following 10 variables.
\describe{ \item{PPTESTCD}{a character vector of CDISC SDTM
PPTESTCD} \item{SYNONYM}{a character vector of CDISC SDTM PPTESTCD
Synonym} \item{NCI}{a character vector of NCI peferred terms}
\item{WNL}{a character vector of WinNonlin(R) software variables}
\item{ExtravascularDefault}{a numeric vector of ordering in report
for extravascular administration, Zero means exclusion in the report.}
\item{ExtravascularWNL}{a numeric vector of WinNonlin(R) style
ordering in report for extravascular administration, Zero means exclusion in
the report.} \item{BolusDefault}{a numeric vector of ordering in
report for extravascular administration, Zero means exclusion in the
report.} \item{BolusWNL}{a numeric vector of WinNonlin(R) style
ordering in report for extravascular administration, Zero means exclusion in
the report.} \item{InfusionDefault}{a numeric vector of ordering in
report for extravascular administration, Zero means exclusion in the
report.} \item{InfusionWNL}{a numeric vector of WinNonlin(R) style
ordering in report for extravascular administration, Zero means exclusion in
the report.} }}
\usage{
data(RptCfg)
}
\description{
Contains the names and order of colum of return table/text by \code{IndiNCA}
and \code{NCA} functions
}
\details{
This table should exist in ncar package. User can edit this table for
shaping the report in one's own style.
}
\keyword{dataset}

