% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCDISC.R
\name{runCDISC}
\alias{runCDISC}
\title{Conduct a noncompartmental analysis with CDISC dataset}
\usage{
runCDISC(wd = getwd(), filenameDM = "DM", filenameEX = "EX",
  filenamePC = "PC", extension = "xpt", incl_arm = NULL)
}
\arguments{
\item{wd}{Working directory of CDISC dataset, containing DM, EX and PC}

\item{filenameDM}{A filename of DM domain. Usually DM or dm}

\item{filenameEX}{A filename of EX domain. Usually EX or ex}

\item{filenamePC}{A filename of PC domain. Usually PC or pc}

\item{extension}{file extension, currently supporting only \code{.xpt}}

\item{incl_arm}{Vector of study arms of interest}
}
\value{
List of output data of noncompartmental analysis
}
\description{
\code{runCDISC} uses ncar package to perform a noncompartmental analysis of
CDISC standardized pharmacokinetic dataset.
}
\examples{
#Currently there is no publicly open CDISC dataset for presenting an example.
}

